Add BaseReactor class and update tests components
