+++ Assets/Scripts/Runtime/ReactiveWorld/Reactor/Tests/TestGlobalReactor.cs
@@ -1,33 +1,9 @@
-using Runtime.ReactiveWorld;
 using Runtime.ReactiveWorld.Reactor;
 using UnityEngine;
 
-public class TestGlobalReactor : MonoBehaviour, IReactor
+public class TestGlobalReactor : BaseReactor
 {
     [SerializeField] private string _reactorName;
-    private WorldManager _worldManager;
 
-    public string Name => _reactorName;
-
-    public bool IsEnabled { get; set; } = true;
-
-    private void Awake()
-    {
-        _worldManager = WorldManager.GetInstance();
-    }
-
-    private void Start()
-    {
-        Initialize();
-    }
-
-    public void Initialize()
-    {
-        _worldManager.Register(this);
-    }
-
-    public void Shutdown()
-    {
-        _worldManager.Unregister(this);
-    }
+    public override string Name => _reactorName;
 }
